api/models/order_discount_test.go
package models

//
//import (
//	"github.com/stretchr/testify/assert"
//	"github.com/vavilen84/gocommerce/constants"
//	"github.com/vavilen84/gocommerce/validator"
//	"testing"
//)
//
//func TestOrderDiscount_ValidateOnCreate(t *testing.T) {
//	m := OrderDiscount{}
//	err := validator.ValidateByScenario(constants.ScenarioCreate, &m, m.getValidator(), m.getValidationRules())
//	assert.NotNil(t, err)
//	assert.NotEmpty(t, err[constants.OrderDiscountOrderIdField])
//	assert.NotEmpty(t, err[constants.OrderDiscountDiscountIdField])
//
//	m = OrderDiscount{
//		OrderId:    1,
//		DiscountId: 1,
//	}
//	err = validator.ValidateByScenario(constants.ScenarioCreate, &m, m.getValidator(), m.getValidationRules())
//	assert.NotNil(t, err)
//}
